final changes to website

+++ scripts/tdd.py
@@ -10,15 +10,18 @@ for snippet in snippets:
         content = f.read()
         code = re.search(codeRe,content).group(1).strip()
         file_to_write_to = open(f'test/{snippet}/{snippet}.py','w')
-        test_file = open(f'test/{snippet}/{snippet}.test.py','w')
         file_to_write_to.write(code)
         file_to_write_to.close()
-        test_file.write(f'''
+        if not os.path.isfile(f'test/{snippet}/{snippet}.test.py'):
+            test_file = open(f'test/{snippet}/{snippet}.test.py','w')
+            test_file.write(f'''
 import types,functools
 from pytape import test
 from {snippet} import {snippet}
 def {snippet}_test(t):
     t.true(isinstance({snippet}, (types.BuiltinFunctionType, types.FunctionType, functools.partial)),'{snippet} is a function')
 test('Testing {snippet}',{snippet}_test)
-        '''.strip())
-        test_file.close()
+            '''.strip())
+            test_file.close()
+        else:
+            pass
